rtcpsu: Fixed MISRAC rule 10.3 violation
This patch fix following violation
"misra_c_2012_rule_10_3_violation:Implicit conversion from
 essential type to different or narrower essential type "signed 32-bit int"

Fixed this violation across the driver by explicitly typecasting at all the
required/intended places.


Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/rtcpsu/src/xrtcpsu.c
@@ -147,7 +147,7 @@ s32 XRtcPsu_CfgInitialize(XRtcPsu *InstancePtr, const XRtcPsu_Config *ConfigPtr,
 	InstancePtr->TimeUpdated = (u32)0U;
 	InstancePtr->CurrTimeUpdated = (u32)0U;
 
-	Status = XST_SUCCESS;
+	Status = (s32)XST_SUCCESS;
 	return Status;
 }
